+++ src/ReferenceToken.php
@@ -47,7 +47,7 @@ final class ReferenceToken
     public static function fromJsonString(string $value): self
     {
         if (1 !== \preg_match('/^(?P<referenceToken>((?P<unescaped>[\x00-\x2E]|[\x30-\x7D]|[\x7F-\x{10FFFF}])|(?P<escaped>~[01]))*)$/u', $value)) {
-            throw Exception\InvalidReferenceToken::fromString($value);
+            throw Exception\InvalidReferenceToken::fromJsonString($value);
         }
 
         return new self($value);