+++ CHANGELOG.md
@@ -14,9 +14,9 @@ For a full diff see [`1.0.0...2.0.0`][1.0.0...2.0.0].
 
 ## Changed
 
-- Renamed named constructors and accessors of `Exception\InvalidJsonPointer`, `JsonPointer`, and `ReferenceToken` ([#4]), by [@localheinz]
+- Renamed named constructors and accessors of `Exception\InvalidJsonPointer`, `JsonPointer`, and `ReferenceToken` ([#4]) and ([#5]), by [@localheinz]
 
-  - `Exception\InvalidJsonPointer::fromString()` to `JsonPointer::fromJsonString()`
+  - `Exception\InvalidJsonPointer::fromString()` to `Exception\InvalidJsonPointer::fromJsonString()`
   - `JsonPointer::fromString()` to `JsonPointer::fromJsonString()`
   - `JsonPointer::toString()` to `JsonPointer::toJsonString()`
   - `ReferenceToken::fromEscapedString()` to `ReferenceToken::fromJsonString()`
@@ -40,5 +40,6 @@ For a full diff see [`a5ba52c...1.0.0`][a5ba52c...1.0.0].
 [#1]: https://github.com/ergebnis/json-pointer/pull/1
 [#2]: https://github.com/ergebnis/json-pointer/pull/2
 [#4]: https://github.com/ergebnis/json-pointer/pull/4
+[#5]: https://github.com/ergebnis/json-pointer/pull/5
 
 [@localheinz]: https://github.com/localheinz