+++ CHANGELOG.md
@@ -38,6 +38,9 @@ For a full diff see [`a5ba52c...1.0.0`][a5ba52c...1.0.0].
 - Added `ReferenceToken` as a value object ([#1]), by [@localheinz]
 - Added `JsonPointer` as a value object ([#2]), by [@localheinz]
 
+[1.0.0]: https://github.com/ergebnis/json-pointer/releases/tag/1.0.0
+[2.0.0]: https://github.com/ergebnis/json-pointer/releases/tag/2.0.0
+
 [a5ba52c...1.0.0]: https://github.com/ergebnis/json-pointer/compare/a5ba52c...1.0.0
 [1.0.0...main]: https://github.com/ergebnis/json-pointer/compare/1.0.0...main
 [2.0.0...main]: https://github.com/ergebnis/json-pointer/compare/2.0.0...main